renamed component, added Add Phase button and form
