+++ packages/webapp/src/components/PhaseForm.scss
@@ -1,23 +1,1 @@
 @import '@fluentui/react/dist/sass/_References';
-
-.phaseDetailRow {
-	display: flex;
-	flex-direction: column;
-	padding: 20px 0;
-    margin-left: 45px;
-    width: calc(100% - 55px);
-
-	.mainRow {
-		display: flex;
-		flex-direction: row;
-
-		.tagDropdown {
-			padding: 0 10px 5px 0;
-		}
-	}
-
-	.detailsRow {
-		display: flex;
-		flex-direction: column;
-	}
-}
